tls: move canonicalIp to internal/tls
