+++ lib/tls.js
@@ -31,6 +31,7 @@ const net = require('net');
 const url = require('url');
 const binding = process.binding('crypto');
 const Buffer = require('buffer').Buffer;
+const canonicalIp = require('internal/tls').canonicalIp;
 
 // Allow {CLIENT_RENEG_LIMIT} client-initiated session renegotiations
 // every {CLIENT_RENEG_WINDOW} seconds. An error event is emitted if more
@@ -164,33 +165,6 @@ function check(hostParts, pattern, wildcards) {
   return true;
 }
 
-exports._canonicalIp = function(address) {
-  // Convert the ip address into the same format
-  // stored in certificates
-  if (net.isIPv6(address)) {
-    const b = ['0', '0', '0', '0', '0', '0', '0', '0'];
-
-    const s = address.split('::');
-    if (s.length === 2) {
-      const s1 = s[0].split(':');
-      for (var n = 0; n < s1.length; n++) {
-        if (s1[n]) {
-          b[n] = s1[n].replace(/^0+(\d+)$/, '$1');
-        }
-      }
-      const s2 = s[1].split(':');
-      for (n = 0; n < s2.length; n++) {
-        if (s2[n]) {
-          b[8 - s2.length + n] = s2[n].replace(/^0+(\d+)$/, '$1');
-        }
-      }
-    }
-
-    return b.join(':');
-  } else
-    return address.replace(/\b0+(\d)/g, '$1'); // Delete leading zeroes
-};
-
 exports.checkServerIdentity = function checkServerIdentity(host, cert) {
   const subject = cert.subject;
   const altNames = cert.subjectaltname;
@@ -217,8 +191,7 @@ exports.checkServerIdentity = function checkServerIdentity(host, cert) {
   let reason = 'Unknown reason';
 
   if (net.isIP(host)) {
-    const canonicalIp = exports._canonicalIp(host);
-    valid = ips.includes(canonicalIp);
+    valid = ips.includes(canonicalIp(host));
     if (!valid)
       reason = `IP: ${host} is not in the cert's list: ${ips.join(', ')}`;
     // TODO(bnoordhuis) Also check URI SANs that are IP addresses.
