+++ node.gyp
@@ -119,6 +119,7 @@
       'lib/internal/readline.js',
       'lib/internal/repl.js',
       'lib/internal/socket_list.js',
+      'lib/internal/tls.js',
       'lib/internal/test/unicode.js',
       'lib/internal/tls.js',
       'lib/internal/url.js',
